pyproject.toml
[tool.poetry]
name = "desafio sn"
version = "0.1.0"
description = "Desafio para contratação de estagio"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-ninja-jwt = "^5.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
